Move animations to walking animation component
