+++ client/src/components/WalkingDog/WalkingDog.js
@@ -9,66 +9,12 @@ import { ReactComponent as RightEye } from '../../assets/animations/walking-dog/
 import { ReactComponent as Tail } from '../../assets/animations/walking-dog/tail.svg';
 import { ReactComponent as Tongue } from '../../assets/animations/walking-dog/tongue.svg';
 
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import './WalkingDog.scss';
 
-function WalkingDog() {
+function WalkingDog({ isHeadBobbing, tongueIsOut, isWaggingTail, isBlinking }) {
 
     let [ isWalking, setIsWalking ] = useState(false);
-
-    let [ isHeadBobbing, setIsHeadBobbing] = useState(false);
-    let [ tongueIsOut, setTongueIsOut ] = useState(false);
-    let [ isWaggingTail, setIsWaggingTail ] = useState(false);
-    let [ isBlinking, setIsBlinking ] = useState(false);
-
-    useEffect(() => {
-        function blink() {
-            function changeEyes() {
-                setIsBlinking(true);
-                setTimeout(
-                    () => setIsBlinking(false),
-                    1000
-                );
-            }
-
-            setTimeout(() => {
-                setInterval(changeEyes, 7000)
-            }, 5000)
-        }
-
-        function wagTail() {
-            function moveTail() {
-                setIsWaggingTail(true);
-                setTimeout(
-                    () => setIsWaggingTail(false), 
-                    2000
-                );
-            }
-
-            setTimeout(() => {
-                setInterval(moveTail, 5000);
-            }, 0);
-        }
-
-        function stickTongueOut() {
-            function changeTongue() {
-                setTongueIsOut(true);
-                setTimeout(
-                    () => setTongueIsOut(false),
-                    1200
-                )
-            }
-
-            setTimeout(() => {
-                setInterval(changeTongue, 15000)
-            }, 7000)
-        }
-
-        blink();
-        wagTail();
-        stickTongueOut();
-
-    }, []);
 
     function determineClasses() {
         let walking = isWalking ? 'walking-dog--isWalking' : null;
